Add throttle util
